requirements.txt
playwright
anthropic
google-api-python-client
google-auth
google-auth-oauthlib
python-dotenv
pandas
requests>=2.33.1
pip-audit
tenacity
beautifulsoup4
# dashboard
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
jinja2>=3.1.4
python-multipart>=0.0.9
sse-starlette>=1.8.2
